Moving to laptop
I fixed the animation timings issue, refactored some enemy and recursive gameflow and overhauled how the flow continues. Began implementing a fix to the size and speed issue. Inside Config, you'll find scaleFactor. Any speed or sizes that are off, edit them to be their original dimension *Config.scaleFactor. I will make a point of doing this bit by bit, but there should be no harm in everyone doing it. scaleFactor is simply a calculation of DPI in relation to a base value I've used, 96, and the actual DPI of the user's screen. I've been developing in and around 96DPI, so it doesn't change much for me, but should for others.